Update lib/registry/mp1.py
A fix to a definite bug in T_s and T_h that caused problems for saturated substances.

A bug still exists that requires a single-valued p to be submitted within a list. Haven't done extensive testing on whether it works with multiple values or not.
Refer to steammp1test.py, lines 184-187 and 200-203