Update steammp1test.py
Weird bug encountered in T_s and T_h when using arrays of s/h that span across phase boundaries. It looks to me like this is some type of issue associated with how you're vectorizing the code and using selective masks. Might overlap with the issue I kludged a fix on p for.